+++ Models/ToolbarStyleSettings.cs
@@ -17,28 +17,4 @@ namespace ExtendedToolbar.Models
         [LocalisableDescription(typeof(ExtendedToolbarStrings), nameof(ExtendedToolbarStrings.SpacerDot))]
         Dot
     }
-
-    /// <summary>
-    /// Акцентные цвета неоновой подсветки тулбара.
-    /// </summary>
-    public enum ToolbarAccentColor
-    {
-        [LocalisableDescription(typeof(ExtendedToolbarStrings), nameof(ExtendedToolbarStrings.AccentPink))]
-        Pink,
-
-        [LocalisableDescription(typeof(ExtendedToolbarStrings), nameof(ExtendedToolbarStrings.AccentPurple))]
-        Purple,
-
-        [LocalisableDescription(typeof(ExtendedToolbarStrings), nameof(ExtendedToolbarStrings.AccentCyan))]
-        Cyan,
-
-        [LocalisableDescription(typeof(ExtendedToolbarStrings), nameof(ExtendedToolbarStrings.AccentLime))]
-        Lime,
-
-        [LocalisableDescription(typeof(ExtendedToolbarStrings), nameof(ExtendedToolbarStrings.AccentGold))]
-        Gold,
-
-        [LocalisableDescription(typeof(ExtendedToolbarStrings), nameof(ExtendedToolbarStrings.AccentWhite))]
-        White
-    }
 }